backend/requirements.txt
Django==5.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
Pillow==10.1.0
python-decouple==3.8
django-cors-headers==4.3.1
django-filter==23.5
drf-spectacular==0.27.0
celery==5.3.4
redis==5.0.1
django-jalali==6.0.1
